pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claas"
version = "0.1.0"
description = "CLaaS: Continual Learning as a Service - SDPO-style distillation API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "CLaaS Team" }
]
dependencies = [
    "modal>=1.0.0",
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "fastapi>=0.110.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "ty>=0.0.1",
]
teacher = [
    "vllm>=0.11.0",
]
eval = [
    "matplotlib>=3.8.0",
    "google-generativeai>=0.5.0",
]

[project.scripts]
claas = "claas.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["claas"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
